+++ app/src/main/java/joshuatee/wx/objects/URL.kt
@@ -27,25 +27,25 @@ class URL(val url: String) {
         return out.toString()
     }
 
-    @Suppress("unused")
-    fun getTextXmlAcceptHeader(): String {
-        UtilityLog.download("getStringFromUrlBaseNoHeader: $url")
-        val breakStr = "ABC123_456ZZ"
-        val out = StringBuilder(5000)
-        try {
-            val request = Request.Builder()
-                .url(url)
-                .header("User-Agent", GlobalVariables.HTTP_USER_AGENT)
-                //.addHeader("Accept", "application/vnd.noaa.dwml+xml;version=1")
-                .addHeader("Accept", "application/atom+xml")
-                .build()
-            val response = MyApplication.httpClient.newCall(request).execute()
-            return response.body.string()
-        } catch (e: Exception) {
-            UtilityLog.handleException(e)
-        }
-        return out.toString().replace(breakStr, "<br>")
-    }
+//    @Suppress("unused")
+//    fun getTextXmlAcceptHeader(): String {
+//        UtilityLog.download("getStringFromUrlBaseNoHeader: $url")
+//        val breakStr = "ABC123_456ZZ"
+//        val out = StringBuilder(5000)
+//        try {
+//            val request = Request.Builder()
+//                .url(url)
+//                .header("User-Agent", GlobalVariables.HTTP_USER_AGENT)
+//                //.addHeader("Accept", "application/vnd.noaa.dwml+xml;version=1")
+//                .addHeader("Accept", "application/atom+xml")
+//                .build()
+//            val response = MyApplication.httpClient.newCall(request).execute()
+//            return response.body.string()
+//        } catch (e: Exception) {
+//            UtilityLog.handleException(e)
+//        }
+//        return out.toString().replace(breakStr, "<br>")
+//    }
 
     fun getBytes(): ByteArray {
         UtilityLog.download("getByte $url")